+++ accounts/forms.py
@@ -1,0 +1,14 @@
+from django.contrib.auth.forms import UserCreationForm
+from .models import CustomUser
+
+class DonorRegistrationForm(UserCreationForm):
+    class Meta(UserCreationForm.Meta):
+        model = CustomUser
+        fields = ('username', 'first_name', 'last_name', 'email')
+
+    def save(self, commit=True):
+        user = super().save(commit=False)
+        user.role = 'donor'
+        if commit:
+            user.save()
+        return user